# PAW-API_Producao/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Informação da RestAPI Produção",
        "version": "1.0.0",
        "title": "PAW - RestAPI Produção",
        "contact": {
            "name": "Grupo - Carlos Pinto, João Pires, Nuno Silva, Sérgio Rodrigues"
        }
    },
    "schemes": [
        "http"
    ],
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/traceabilitypurchases": {
            "get": {
                "summary": "Obter toda a rastreabilidade de items comprados",
                "description": "Obter toda a rastreabilidade de items comprados",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "operação com sucesso",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/TraceabilityPurchase"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            },
            "post": {
                "summary": "Criar novo rastreio de item comprado",
                "description": "Criar novo rastreio de item comprado",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "rastreio de compra",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "type": "object",
                                "$ref": "#/definitions/TraceabilityPurchase"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/TraceabilityPurchase"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/traceabilitypurchases/traceabilitypurchase?page={nPaginas}&limit={limite}": {
            "get": {
                "summary": "Obter toda a rastreabilidade de items comprados, ordenada por nº de lote, definindo nº de páginas e limite",
                "description": "Obter toda a rastreabilidade de items comprados, ordenada por nº de lote, definindo nº de páginas e limite",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "nPaginas",
                        "description": "Número de páginas"
                    },
                    {
                        "in": "path",
                        "name": "limite",
                        "description": "Limite de rastreabilidade de items comprados a apresentar"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/TraceabilityPurchase"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/traceabilitypurchases/{traceId}": {
            "get": {
                "summary": "Obter rastreio de item comprado por id",
                "description": "Obter rastreio de item comprado por id",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "traceId",
                        "description": "id do rastreio a pesquisar",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "task": {
                                    "type": "object",
                                    "$ref": "#/definitions/TraceabilityPurchase"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/TraceabilityPurchase"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            },
            "patch": {
                "summary": "Atualizar rastreio de item comprado",
                "description": "Atualizar rastreio de item comprado",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "traceId",
                        "in": "path",
                        "description": "id do rastreio a alterar",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "rastreio a alterar",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "type": "object",
                                "$ref": "#/definitions/TraceabilityPurchase"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/TraceabilityPurchase"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Apagar rastreio de item comprado",
                "description": "Apagar rastreio de item comprado",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "traceId",
                        "in": "path",
                        "description": "id do rastreio a eliminar",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/TraceabilityPurchase"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
...
        "/search/traceabilitypurchase?tags={querie}": {
            "get": {
                "summary": "Obter rastreabilidade de items comprados através de pesquisa",
                "description": "Obter rastreabilidade de items comprados através de pesquisa",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "querie",
                        "description": "termo a pesquisar",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "task": {
                                    "type": "object",
                                    "$ref": "#/definitions/TraceabilityPurchase"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/TraceabilityPurchase"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "TraceabilityPurchase": {
            "type": "object",
            "properties": {
                "batch_number": {
                    "type": "string"
                },
                "expiration_date": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "purchaseorder_id": {
                    "type": "string"
                },
                "item_id": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "InvalidResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}